Fixed an issue of version detection for the application
Version specified inside Info.plist is just a generic string "{APP_VERSION}" which is the same in every version of the application. Added URL search to retrieve the version number from the download URL. This means that the munki recipe will not be able to correctly detect the version for this application so I'm removing this recipe for now.